+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. This change
 - Command-line option to control the number of rows generated for each table.
 - Support for date / datetime / timestamp generators.
 - Improved number type recognition so that types like BIGINT and DOUBLE would be recognized.
+- Added headers to CSV output.
 
 ## [0.2.0] - 2017-08-17
 ### Added