+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file. This change
 - Added random generators to select first and last names from U.S. Census Data.
 - Added composite generator for full names.
 - Added `--list-generators` command-line option so that the user can see what generators are provided.
+- Added command line option (`-s`) to limit the schemas being scanned. By default, all schemas are scanned but provision of a list of schemas will limit scans to those.
 
 ## [0.2.0] - 2017-08-17
 ### Added