+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/). This project follows the conventions of [semver.org](http://semver.org/).
 
-## [Unreleased]
+## [0.3.0] - 2017-08-24
 ### Added
 - Command-line option to control the number of rows generated for each table.
 - Support for date / datetime / timestamp generators.